Filling in missing parts of transaction
working in master now there were a few things of my work that was missing so I am adding them back in. sorry for the back to back commits. Also started on filling in the test cases for the product object.

Plan on finishing this and moving onto the transaction tests.